requirements.txt
prometheus_client
speedtest-cli